requirements.txt
streamlit>=1.34.0
torch>=2.2.0
diffusers>=0.27.2
transformers>=4.41.0
accelerate>=0.30.0
moviepy>=1.0.3
tqdm>=4.66.4
Pillow>=10.3.0
